backend/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

# Image processing and OCR
opencv-python>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Validation
pydantic>=2.5.0

# HTTP client
httpx>=0.25.0

# Optional: ML libraries (comment out if causing issues)
# tensorflow>=2.13.0
# torch>=2.0.0
# torchvision>=0.15.0
# scikit-learn>=1.3.0 